% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margarita_volume.R
\name{margarita}
\alias{margarita}
\title{Margarita Recipe}
\usage{
margarita(margarita_volume)
}
\arguments{
\item{margarita_volume}{Volume of margarita in milliliters.}
}
\value{
A named list containing the amounts of lime, cointreau, tequila, and syrup.
}
\description{
Calculate the ingredients needed for a margarita based on the volume.
}
\examples{
margarita(260)
}
